Fix maven report background thread exception and update logs
